ensure changes to the physical planner are reflected here
